+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q035/Main.java
@@ -1,8 +1,6 @@
 package com.noob.algorithm.leetcode.q035;
 
 
-import com.noob.algorithm.leetcode.q279.Solution;
-
 import java.util.ArrayList;
 import java.util.List;
