+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q279/Main.java
@@ -1,5 +1,8 @@
-package com.noob.algorithm.leetcode.q152;
+package com.noob.algorithm.leetcode.q279;
 
+
+import com.noob.algorithm.base.util.PrintListUtil;
+import com.noob.algorithm.leetcode.q019.Solution2;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,22 +10,22 @@ import java.util.List;
 public class Main {
 
     // 初始化测试用例
-    public static List<int[]> initCase() {
-        List<int[]> caseList = new ArrayList<>();
+    public static List<Integer> initCase() {
+        List<Integer> caseList = new ArrayList<>();
 
         // todo 补充测试用例集合
-        caseList.add(new int[]{2, 3, -2, 4}); // 6
-        caseList.add(new int[]{-2, 0, -1}); // 0
+        caseList.add(12);
+        caseList.add(13);
 
         return caseList;
     }
 
-    public static void testSolution1(List<int[]> caseList) {
+    public static void testSolution1(List<Integer> caseList) {
 
-        for (int[] nums : caseList) {
+        for (int n : caseList) {
             // ① 调用算法
             Solution1 solution = new Solution1();
-            int res = solution.maxProduct(nums);
+            int res = solution.numSquares(n);
 
             // ② 打印数据
             System.out.println(res);
@@ -35,7 +38,7 @@ public class Main {
     }
 
 
-    public static void testSolution2(List<int[]> caseList) {
+    public static void testSolution2(List<Integer> caseList) {
 
     }
 
@@ -43,7 +46,7 @@ public class Main {
     public static void main(String[] args) {
 
         // 参数初始化
-        List<int[]> caseList = initCase();
+        List<Integer> caseList = initCase();
 
         System.out.println("---------- 算法①调用测试 ----------");
